Removed info on bitpay's library, updated
Bitpay has a library called 'bitcore' which is used to access the bitcoin blockchain. This is unrelated to the Bitcoin coin, which was developed by a different group. The previous file version conflated the two, resulting in confusion. This material was taken out in this version, and updated material was added to reflect spec changes and new developments.