BTX Coresight API info added
added the correct API to use for merchants to make calls to the Bitcore blockchain. This is a later version of the original library mentioned in an earlier version of this page, but which has been modified appropriately for Bitcore. Also added corresponding references to it in the table of contents, and added the wallet based on it to the list of wallets.